+++ .build-banner.js
@@ -1,5 +1,0 @@
-window.__ModuleLoader__.load({
-  id: "meow-memory",
-  factory: (require) => {
-    var module = { exports: {} };
-    var exports = module.exports;